backend/requirements.txt
email-validator==2.2.0
dnspython==2.7.0
elasticsearch==8.13.0
fastapi==0.111.0
motor==3.4.0
numpy==1.26.4
openai==1.109.1
pandas==2.2.2
passlib==1.7.4
pdf2image==1.17.0
pillow==10.3.0
plaid-python==35.0.0
pydantic==2.11.7
pymongo==4.6.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.31.0
starlette==0.37.2
stripe==12.4.0
torch==2.8.0
user-agents==2.2.0
transformers==4.55.2
uvicorn[standard]==0.29.0
google-auth==2.41.1
attrs==25.3.0
boto3==1.40.39
cryptography==46.0.3
tiktoken==0.12.0
croniter==6.0.0
intuit-oauth==1.2.6
pyotp==2.9.0
qrcode==8.2
